Cleaned up Model-View relationships. Bonding is done. Some UI changes.
